refactor(runtime): move js_proxy_set/proxy_set_with_receiver into proxy/put_value.rs
proxy.rs grew to 2004 lines after rebasing onto main (which independently
added ~30 lines), tripping the 2000-line file-size CI gate. Relocate the
Proxy [[Set]] entry points into the sibling put_value.rs submodule
(already the natural home for PutValue/[[Set]] dispatch) rather than
trim comments to fit — same pattern as the file's other topical splits.
Pure code motion, no behavior change (re-verified: built-ins/Proxy still
219 pass / 21 runtime-fail).
